requirements.txt
Flask==2.3.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
pandas==2.1.4
mlxtend==0.22.0
Pillow==10.2.0
numpy==1.24.4
scikit-learn==1.3.0
tqdm==4.66.1

torch
torchvision
gdown

google-cloud-storage>=2.0.0
google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0   # optional, only if you do OAuth
python-dotenv==1.0.0
werkzeug==2.3.6
